refactor: BaseballNumber util package로 이동
