+++ src/main/java/baseball/Application.java
@@ -1,8 +1,8 @@
 package baseball;
 
 import baseball.controller.BaseballController;
-import baseball.service.BaseballNumberGenerator;
 import baseball.service.BaseballService;
+import baseball.util.BaseballNumberGenerator;
 import baseball.view.InputView;
 import baseball.view.OutputView;
 
@@ -12,7 +12,8 @@ public class Application {
         OutputView outputView = new OutputView();
         BaseballNumberGenerator baseballNumberGenerator = new BaseballNumberGenerator();
         BaseballService baseballService = new BaseballService();
-        BaseballController baseballController = new BaseballController(inputView, outputView, baseballNumberGenerator, baseballService);
+        BaseballController baseballController = new BaseballController(inputView, outputView, baseballNumberGenerator,
+                baseballService);
 
         baseballController.startBaseballGame();
