+++ src/main/java/baseball/util/BaseballNumberGenerator.java
@@ -1,7 +1,6 @@
-package baseball.service;
+package baseball.util;
 
 import camp.nextstep.edu.missionutils.Randoms;
-
 import java.util.ArrayList;
 import java.util.List;
